+++ src/Base/Traits/ApplicationTrait.php
@@ -14,6 +14,7 @@ use Hirtz\Skeleton\Db\ActiveQuery;
 use Hirtz\Skeleton\Db\DatabaseComponents;
 use Hirtz\Skeleton\Models\Collections\TrailModelCollection;
 use Hirtz\Skeleton\Models\Definitions\DefinitionRegistry;
+use Hirtz\Skeleton\Models\Definitions\Definitions;
 use Hirtz\Skeleton\Models\User;
 use Hirtz\Skeleton\Modules\Admin\Module;
 use Hirtz\Skeleton\Rbac\DbManager;
@@ -94,6 +95,9 @@ trait ApplicationTrait
                 ],
                 'cache' => [
                     'class' => FileCache::class,
+                ],
+                'definitions' => [
+                    'class' => Definitions::class,
                 ],
                 'db' => [
                     'class' => Connection::class,